requirements.txt
# Core dependencies
openai>=1.0.0
pinecone>=5.0.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Document processing
python-docx>=1.0.0
PyPDF2>=3.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
